+++ src/main/java/traben/entity_model_features/EMFManager.java
@@ -24,6 +24,7 @@ import traben.entity_model_features.models.animation.math.variables.EMFModelOrRe
 import traben.entity_model_features.models.jem_objects.EMFJemData;
 import traben.entity_model_features.models.parts.EMFModelPartVanilla;
 import traben.entity_model_features.utils.EMFDirectoryHandler;
+import traben.entity_model_features.utils.EMFResourceCaching;
 import traben.entity_model_features.utils.EMFUtils;
 import traben.entity_texture_features.ETF;
 import traben.entity_texture_features.utils.ETFLruCache;
@@ -115,7 +116,7 @@ public class EMFManager {//singleton for data holding and resetting needs
         EMFUtils.log("Clearing data for reload.", false);
         EMF.config().loadFromFile();
         EMFModelMappings.UNKNOWN_MODEL_MAP_CACHE.clear();
-        EMFDirectoryHandler.clearCache();
+        EMFResourceCaching.clearCache();
         self = new EMFManager();
     }
 
@@ -129,7 +130,7 @@ public class EMFManager {//singleton for data holding and resetting needs
         try {
             var jemLoc = EMFUtils.res(pathOfJem);
             var resourceManager = Minecraft.getInstance().getResourceManager();
-            if (!EMFDirectoryHandler.resourceExists(resourceManager, jemLoc)) {
+            if (!EMFResourceCaching.resourceExists(resourceManager, jemLoc)) {
                 if (print) EMFUtils.log(pathOfJem + ", .jem read failed " + pathOfJem + " does not exist", false);
                 return null;
             }
